rename lib-franklin.js to aem.js
