+++ 404.html
@@ -14,7 +14,7 @@
   <link rel="icon" type="image/svg+xml" href="/resources/img/adaptto-favicon.svg">
   <script src="/scripts/scripts.js" type="module" crossorigin="use-credentials"></script>
   <script type="module">
-    import { sampleRUM } from '/scripts/lib-franklin.js';
+    import { sampleRUM } from '/scripts/aem.js';
 
     window.addEventListener('load', () => {
       if (document.referrer) {
