+++ blocks/speaker-gallery/speaker-gallery.js
@@ -1,4 +1,4 @@
-import { createOptimizedPicture, readBlockConfig } from '../../scripts/lib-franklin.js';
+import { createOptimizedPicture, readBlockConfig } from '../../scripts/aem.js';
 import { getQueryIndex } from '../../scripts/services/QueryIndex.js';
 import { append } from '../../scripts/utils/dom.js';
 import { parseCSVArray } from '../../scripts/utils/metadata.js';
