+++ head.html
@@ -1,5 +1,5 @@
 <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=2"/>
-<script src="/scripts/lib-franklin.js" type="module"></script>
+<script src="/scripts/aem.js" type="module"></script>
 <script src="/scripts/scripts.js" type="module"></script>
 <link rel="stylesheet" href="/styles/styles.css">
 <link rel="shortcut icon" href="/resources/img/adaptto-favicon.ico"/>
