+++ scripts/delayed.js
@@ -1,5 +1,5 @@
 // eslint-disable-next-line import/no-cycle
-import { sampleRUM } from './lib-franklin.js';
+import { sampleRUM } from './aem.js';
 import { decorateConsentManagement } from './utils/usercentrics.js';
 
 // enable UserCentrics consent management
