+++ sonar-project.properties
@@ -2,4 +2,4 @@ sonar.organization=adaptto
 sonar.projectKey=adaptto_adaptto-website
 sonar.javascript.lcov.reportPaths=./coverage/lcov.info
 sonar.coverage.exclusions=test/**
-sonar.exclusions=scripts/lib-franklin.js,scripts/3rdparty/**
+sonar.exclusions=scripts/aem.js,scripts/3rdparty/**
