+++ test/scripts/scripts.test.js
@@ -13,7 +13,7 @@ document.head.innerHTML = await readFile({ path: './head.html' });
 
 describe('Core Helix features', () => {
   before(async () => {
-    lib = await import('../../scripts/lib-franklin.js');
+    lib = await import('../../scripts/aem.js');
 
     document.body.innerHTML = await readFile({ path: './body.html' });
   });
